Console with command pattern (#40)
* Update 2d renderer to render from window coordinates

* Implement basic command pattern

Add project related commands

* Create spearate console project

Separate the console class form the main project

Add mechanisms for registering custom commands

Add mechanisms for runtime creation of the command objects

Implement automatic command and argument resolution

Replace the old console with the new one

Add tests for the new console

* Fix crash on empty commands

* Fix rectangle rendering after text rendering

+++ src/graphics/font.hpp
@@ -14,7 +14,7 @@ public:
         // offset from baseline to left/top of glyph
         glm::uvec2 _bearing;
         // offset to advance to next glyph
-        long _advance;
+        float _advance;
     };
 
 public:
